+++ CHANGELOG.md
@@ -8,9 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Codelist support for all versions (i.e. v1.01+)
+- Add `start` and `end` properties to activities
 
 ### Changed
 - Change internal representation of codelists
+- Return python `date` objects for dates
 
 ## [1.3.1] – 2018-12-17